Clean up all CI warnings
Silence the MSVC conversion warnings with explicit casts, cast the
bool static assert trick in blake2.h to int, and drop the leftover
avg hashrate scaffolding in the speed printout. CI moves checkout and
upload-artifact to v7 for the node 20 deprecation, drops the
msvc-dev-cmd action since the Visual Studio generator and vcpkg find
MSVC on their own, and untaps aws/tap so homebrew stops complaining.

+++ src/backend/common/WorkerJob.h
@@ -86,7 +86,7 @@ public:
     }
 
 
-    inline int32_t nonceOffset() const { return currentJob().nonceOffset(); }
+    inline int32_t nonceOffset() const { return static_cast<int32_t>(currentJob().nonceOffset()); }
     inline size_t nonceSize() const { return currentJob().nonceSize(); }
 
 private:
